Handle stopping containers gracefully
There is a race condition between getting the list of containers
belonging to a Docker Compose project and fetching the corresponding
IPs for each container, since a container can stop (and consequently
lose its IP).

We can deal with this race by making the code that fetches the IPs
expect stopped containers.